infrastructure/gateway/requirements.txt
# AIVillage Unified API Gateway Requirements
# Production-ready dependencies for the unified API gateway

# Core FastAPI stack
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Authentication and Security
PyJWT>=2.8.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Async and WebSocket support
websockets>=12.0
aiofiles>=23.0.0
httpx>=0.25.0

# Middleware and utilities
slowapi>=0.1.9  # Rate limiting
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
isort>=5.12.0
mypy>=1.7.0

# Optional AI/ML dependencies (for Agent Forge integration)
torch>=2.0.0; python_version>="3.8"
transformers>=4.35.0; python_version>="3.8"
datasets>=2.14.0; python_version>="3.8"

# Optional P2P/Fog dependencies
libp2p>=0.1.0; python_version>="3.8"
aioquic>=0.9.0; python_version>="3.8"
